fix: authZ checks for delete and update process
(cherry picked from commit 49eca561927f4e46b053b0fe01e52c2e20dd4b50)

+++ dspace-api/src/main/java/org/dspace/scripts/ProcessServiceImpl.java
@@ -142,7 +142,7 @@ public class ProcessServiceImpl implements ProcessService {
     }
 
     @Override
-    public void start(Context context, Process process) throws SQLException {
+    public void start(Context context, Process process) throws SQLException, AuthorizeException {
         process.setProcessStatus(ProcessStatus.RUNNING);
         process.setStartTime(new Date());
         update(context, process);
@@ -152,7 +152,7 @@ public class ProcessServiceImpl implements ProcessService {
     }
 
     @Override
-    public void fail(Context context, Process process) throws SQLException {
+    public void fail(Context context, Process process) throws SQLException, AuthorizeException {
         process.setProcessStatus(ProcessStatus.FAILED);
         process.setFinishedTime(new Date());
         update(context, process);
@@ -162,7 +162,7 @@ public class ProcessServiceImpl implements ProcessService {
     }
 
     @Override
-    public void complete(Context context, Process process) throws SQLException {
+    public void complete(Context context, Process process) throws SQLException, AuthorizeException {
         process.setProcessStatus(ProcessStatus.COMPLETED);
         process.setFinishedTime(new Date());
         update(context, process);
@@ -174,6 +174,9 @@ public class ProcessServiceImpl implements ProcessService {
     @Override
     public void appendFile(Context context, Process process, InputStream is, String type, String fileName)
         throws IOException, SQLException, AuthorizeException {
+        if (!authorizeActionBoolean(context, process)) {
+            throw new AuthorizeException("Cannot append file to process " + process.getID());
+        }
         Bitstream bitstream = bitstreamService.create(context, is);
         if (getBitstream(context, process, type) != null) {
             throw new IllegalArgumentException("Cannot create another file of type: " + type + " for this process" +
@@ -194,7 +197,9 @@ public class ProcessServiceImpl implements ProcessService {
 
     @Override
     public void delete(Context context, Process process) throws SQLException, IOException, AuthorizeException {
-
+        if (!authorizeActionBoolean(context, process)) {
+            throw new AuthorizeException("Cannot delete process " + process.getID());
+        }
         for (Bitstream bitstream : ListUtils.emptyIfNull(process.getBitstreams())) {
             bitstreamService.delete(context, bitstream);
         }
@@ -204,7 +209,10 @@ public class ProcessServiceImpl implements ProcessService {
     }
 
     @Override
-    public void update(Context context, Process process) throws SQLException {
+    public void update(Context context, Process process) throws SQLException, AuthorizeException {
+        if (!authorizeActionBoolean(context, process)) {
+            throw new AuthorizeException("Cannot update process " + process.getID());
+        }
         processDAO.save(context, process);
     }
 
@@ -329,6 +337,30 @@ public class ProcessServiceImpl implements ProcessService {
     @Override
     public int countByUser(Context context, EPerson user) throws SQLException {
         return processDAO.countByUser(context, user);
+    }
+
+    /**
+     * Authorize any action, to ensure only a process creator/owner or a repository administrator
+     * may read, update or delete an existing process.
+     * @param context DSpace context containing the current user
+     * @param process the process to check
+     * @return true if the current user may perform the action, or false
+     */
+    @Override
+    public boolean authorizeActionBoolean(Context context, Process process) {
+        try {
+            if (process == null) {
+                return false;
+            }
+            // Only the process owner or an administrator may perform any action
+            if ((null != context.getCurrentUser() && context.getCurrentUser().equals(process.getEPerson()) )
+                    || authorizeService.isAdmin(context)) {
+                return true;
+            }
+        } catch (SQLException e) {
+            log.error(e::getMessage, e);
+        }
+        return false;
     }
 
     @Override
